frontend/webapp/requirements.txt
pymongo
flask
PyYAML
